feat(db-client): add a file picker for the sqlite database path
The SQLite driver takes its database as a filesystem path, but the
connection form only offered a free-text input — the user had to know
and type the absolute path with no way to browse for it. A filesystem
browser already existed, but it was welded to project selection:
folder-only, project badges hardcoded, title and confirm label fixed,
so it could not be reused as-is.

Promote FolderBrowser's body into a generic PathBrowser that takes the
variable parts as props: directory or file mode, title and confirm
label, an extension filter with an "All files" escape hatch (a SQLite
database may carry no extension at all), an initial path that opens a
file's parent with the file preselected, and opt-in project badges and
folder actions. FolderBrowser becomes a thin project-folder preset over
it, so both navigators keep working unchanged, and the connection form
mounts the same component in file mode behind a browse button next to
the Database field.

The form also offered an SSH Tunnel section for SQLite. A tunnel
forwards a TCP port and SQLite has none, so the backend already skipped
it for this driver — the section was dead UI that could still block a
save, because enabling it demanded a host and username that were never
going to be used. Gate it, the host/credential block, and the SSH
validation on one `isNetworkDriver` flag.

Three layout issues surfaced while using the generalized picker, fixed
in the shared component so the project picker benefits too: the footer
held a single button against an otherwise empty bar, the listing stayed
one column no matter how wide the modal got, and the active toolbar
pill used a violet-900/30 wash that was indistinguishable from the
slate toolbar behind it in dark mode.
