+++ frontend/components/db-client/sidebar/ConnectionForm.svelte
@@ -1,6 +1,7 @@
 <script lang="ts">
 	import { untrack } from 'svelte';
 	import Icon from '$frontend/components/common/display/Icon.svelte';
+	import PathBrowser from '$frontend/components/common/form/PathBrowser.svelte';
 	import { dbClientStore } from '$frontend/stores/features/db-client.svelte';
 	import type {
 		DbClientConnection,
@@ -26,6 +27,9 @@
 		mssql: 1433,
 		sqlite: null
 	};
+
+	// Extensions the SQLite file picker filters on — it can still show every file
+	const SQLITE_EXTENSIONS = ['db', 'sqlite', 'sqlite3', 'db3'];
 
 	const DRIVER_LABELS: Record<DbDriver, string> = {
 		mysql: 'MySQL',
@@ -55,13 +59,18 @@
 	let sshPrivateKey = $state(initial?.ssh.privateKey ?? '');
 	let sshPassphrase = $state(initial?.ssh.passphrase ?? '');
 
+	let dbFilePickerOpen = $state(false);
+
 	let testing = $state(false);
 	let saving = $state(false);
 	let testResult = $state<DbClientHealth | null>(null);
 	let formError = $state<string | null>(null);
 
 	const isSqlite = $derived(driver === 'sqlite');
-	const showHostFields = $derived(!isSqlite);
+	// SQLite opens a file on the server's filesystem: no host, no port, no
+	// credentials, and nothing for an SSH tunnel to forward — the backend skips
+	// the tunnel for this driver outright (connection-manager openEntry).
+	const isNetworkDriver = $derived(!isSqlite);
 
 	const hostPlaceholder = '127.0.0.1';
 	const portPlaceholder = $derived(DEFAULT_PORTS[driver]?.toString() ?? '');
@@ -83,7 +92,7 @@
 	}
 
 	function buildInput(): DbClientConnectionInput {
-		const ssh = sshEnabled
+		const ssh = isNetworkDriver && sshEnabled
 			? {
 					enabled: true,
 					host: sshHost,
@@ -96,8 +105,8 @@
 				}
 			: { enabled: false };
 
-		const resolvedHost = showHostFields ? (host.trim() || hostPlaceholder) : undefined;
-		const resolvedPort = showHostFields ? (port ?? DEFAULT_PORTS[driver] ?? undefined) : undefined;
+		const resolvedHost = isNetworkDriver ? (host.trim() || hostPlaceholder) : undefined;
+		const resolvedPort = isNetworkDriver ? (port ?? DEFAULT_PORTS[driver] ?? undefined) : undefined;
 
 		return {
 			name: name.trim(),
@@ -114,7 +123,7 @@
 	function validate(): string | null {
 		if (!name.trim()) return 'Name is required';
 		if (driver === 'sqlite' && !database.trim()) return 'SQLite needs a file path in Database field';
-		if (sshEnabled) {
+		if (isNetworkDriver && sshEnabled) {
 			if (!sshHost.trim()) return 'SSH host is required';
 			if (!sshUsername.trim()) return 'SSH username is required';
 			if (sshAuthMethod === 'password' && !sshPassword) return 'SSH password is required';
@@ -193,7 +202,7 @@
 		</select>
 	</label>
 
-	{#if showHostFields}
+	{#if isNetworkDriver}
 		<div class="grid grid-cols-3 gap-2">
 			<label class="col-span-2 flex flex-col gap-1">
 				<span class="text-xs text-slate-500 dark:text-slate-400">Host</span>
@@ -248,15 +257,47 @@
 				<span class="text-slate-400">(file path)</span>
 			{/if}
 		</span>
-		<input
-			type="text"
-			bind:value={database}
-			placeholder={isSqlite ? '/absolute/path/to.db' : ''}
-			class="px-2.5 py-1.5 bg-white dark:bg-slate-900 border border-slate-200 dark:border-slate-700 rounded-md text-sm text-slate-900 dark:text-slate-100"
-		/>
+		<div class="flex items-center gap-2">
+			<input
+				type="text"
+				bind:value={database}
+				placeholder={isSqlite ? '/absolute/path/to.db' : ''}
+				class="flex-1 min-w-0 px-2.5 py-1.5 bg-white dark:bg-slate-900 border border-slate-200 dark:border-slate-700 rounded-md text-sm text-slate-900 dark:text-slate-100"
+			/>
+			{#if isSqlite}
+				<button
+					type="button"
+					class="flex shrink-0 p-2 rounded-md border border-slate-200 dark:border-slate-700 text-slate-600 dark:text-slate-300 hover:bg-slate-100 dark:hover:bg-slate-800"
+					onclick={() => (dbFilePickerOpen = true)}
+					title="Browse for a database file"
+					aria-label="Browse for a database file"
+				>
+					<Icon name="lucide:folder-open" class="w-4 h-4" />
+				</button>
+			{/if}
+		</div>
 	</label>
 
-	<!-- SSH section -->
+	{#if isSqlite}
+		<PathBrowser
+			bind:isOpen={dbFilePickerOpen}
+			mode="file"
+			title="Select Database File"
+			confirmText="Use File"
+			fileExtensions={SQLITE_EXTENSIONS}
+			initialPath={database.trim()}
+			allowFolderActions={false}
+			onClose={() => (dbFilePickerOpen = false)}
+			onSelect={(path) => {
+				database = path;
+				dbFilePickerOpen = false;
+			}}
+		/>
+	{/if}
+
+	<!-- SSH section — network drivers only; a tunnel forwards a TCP port and
+	     SQLite has none -->
+	{#if isNetworkDriver}
 	<div class="border-t border-slate-200 dark:border-slate-800 pt-3">
 		<button
 			type="button"
@@ -353,6 +394,7 @@
 			</div>
 		{/if}
 	</div>
+	{/if}
 
 	<!-- Test result -->
 	{#if testResult}
